WR14532 Clean up static library dependencies.
New wxthings libs for SunOS (including debug versions -- not that we use them anywhere, though). wxthings was modified to have some of the wxCustomButton changes that we'd made when we were using wxCustomButton from wxSDL. Added some Makefiles that work.

Owners reviewed: no, because some days I just don't care. And what are the co-owners going to review, anyway? I've introduced some Makefiles that actually *build* wxthings and wxplotctrl. And I modified the top-level Contrib Makefile to call the Makefiles that actually build wxthings and wxplotctrl. And comparing .a (lib) files is pointless.

git-svn-id: https://isrsvn.usurf.usu.edu/svn/isr/trunk/Contrib/wxthings@9060 a06e8daa-702c-0410-a579-eb5e6c4fa09d